+++ src/components/Sidebar/index.tsx
@@ -1,0 +1,25 @@
+import { Box, Button, Typography } from '@mui/material';
+
+export const Sidebar = () => {
+	return (
+		<Box
+			sx={{
+				width: '200px',
+				display: 'flex',
+				flexDirection: 'column',
+				justifyContent: 'space-between',
+			}}
+		>
+			<Box>
+				<Typography variant="h5" sx={{ padding: '10px' }}>
+					ChatApp
+				</Typography>
+			</Box>
+			<Box>
+				<Button variant="outlined" sx={{ width: '100%' }}>
+					ログアウト
+				</Button>
+			</Box>
+		</Box>
+	);
+};
