+++ src/features/chat/components/page/ChatPage/index.tsx
@@ -1,0 +1,12 @@
+import { Layout } from "@/components/Layout"
+import { Chat } from "../../ui/Chat"
+
+const ChatPage = () => {
+	return (
+		<Layout>
+			<Chat />
+		</Layout>
+	)
+}
+
+export default ChatPage
